MetaGME 64bit
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1613 b35830b6-564f-0410-ac5f-b387e50cb686

+++ SDK/DecoratorLib/Dibapi.cpp
@@ -473,7 +473,7 @@ HGLOBAL WINAPI CopyHandle (HGLOBAL h)
 	if (h == NULL)
 		return NULL;
 
-	DWORD dwLen = ::GlobalSize((HGLOBAL) h);
+	SIZE_T dwLen = ::GlobalSize((HGLOBAL) h);
 	HGLOBAL hCopy = ::GlobalAlloc(GHND, dwLen);
 
 	if (hCopy != NULL)
